+++ apps/web/pages/settings/organizations/[id]/about.tsx
@@ -1,14 +1,12 @@
-import type { GetServerSidePropsContext } from "next";
 import { useRouter } from "next/router";
 
-import { getServerSession } from "@calcom/features/auth/lib/getServerSession";
 import { AboutOrganizationForm } from "@calcom/features/ee/organizations/components";
-import { getFeatureFlagMap } from "@calcom/features/flags/server/utils";
 import { useLocale } from "@calcom/lib/hooks/useLocale";
-import { MembershipRole } from "@calcom/prisma/client";
 import { WizardLayout, Meta } from "@calcom/ui";
 
 import PageWrapper from "@components/PageWrapper";
+
+export { getServerSideProps } from "@calcom/features/ee/organizations/pages/organization";
 
 const AboutOrganizationPage = () => {
   const { t } = useLocale();
@@ -27,44 +25,6 @@ const LayoutWrapper = (page: React.ReactElement) => {
       {page}
     </WizardLayout>
   );
-};
-
-export const getServerSideProps = async ({ req, res }: GetServerSidePropsContext) => {
-  const prisma = await import("@calcom/prisma").then((mod) => mod.default);
-  const flags = await getFeatureFlagMap(prisma);
-  // Check if organizations are enabled
-  if (flags["organizations"] !== true) {
-    return {
-      notFound: true,
-    };
-  }
-
-  // Check if logged in user has an organization assigned
-  const session = await getServerSession({ req, res });
-  if (!session?.user.organizationId) {
-    return {
-      notFound: true,
-    };
-  }
-  // Check if logged in user has OWNER/ADMIN role in organization
-  const membership = await prisma.membership.findFirst({
-    where: {
-      userId: session?.user.id,
-      teamId: session?.user.organizationId,
-    },
-    select: {
-      role: true,
-    },
-  });
-  if (!membership?.role || membership?.role === MembershipRole.MEMBER) {
-    return {
-      notFound: true,
-    };
-  }
-
-  return {
-    props: {},
-  };
 };
 
 AboutOrganizationPage.getLayout = LayoutWrapper;
